Split to segment.cpp, move closestPointOnSeg, move segment length, delete isDegenerate
