+++ src/software/new_geom/segment.cpp
@@ -20,33 +20,12 @@ const Point& Segment::getEnd() const
     return end;
 }
 
-inline Segment::Segment() {}
-
-inline Segment::Segment(const Point& start, const Point& end) : start(start), end(end)
+double Segment::length()
 {
+    return (getStart() - getEnd()).length();
 }
 
-inline Segment Segment::reverse() const
+double Segment::lengthSquared()
 {
-    return Segment(end, start);
-}
-
-inline Vector Segment::toVector() const
-{
-    return end - start;
-}
-
-inline double Segment::length() const
-{
-    return (end - start).length();
-}
-
-inline double Segment::slope() const
-{
-    return (end.y() - start.y()) / (end.x() - start.x());
-}
-
-inline bool Segment::operator==(const Segment& other) const
-{
-    return start == other.start && end == other.end;
+    return (getStart() - getEnd()).lengthSquared();
 }
